+++ sample/SampleService.Kestrel/Controllers/ValuesController.cs
@@ -1,8 +1,12 @@
 ﻿using System.Collections.Generic;
+using Butterfly.Client;
 using Microsoft.AspNetCore.Mvc;
 
 namespace SampleService.Kestrel.Controllers
 {
+    /// <summary>
+    /// 测试API
+    /// </summary>
     [Route("api/[controller]")]
     public class ValuesController : Controller
     {
@@ -11,7 +15,7 @@ namespace SampleService.Kestrel.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        public IEnumerable<string> Get()
+        public IEnumerable<string> Get([FromServices] IServiceTracer tracer)
         {
             return new[] { "value1", "value2" };
         }
