+++ sample/SampleService.Kestrel/appsettings.json
@@ -15,5 +15,8 @@
   "AppSettings": {
     "VirtualDirectory": "/Api",
     "IdentityServerAuthority": "http://localhost:5000"
+  },
+  "Butterfly": {
+    "CollectorUrl": "http://localhost:9618"
   }
 }
