+++ src/NanoFabric.Ocelot/NanoFabric.Ocelot.csproj
@@ -15,8 +15,9 @@
     <PackageReference Include="App.Metrics.AspNetCore.Reporting" Version="2.0.0-alpha" />
     <PackageReference Include="App.Metrics.AspNetCore.Tracking" Version="2.0.0-alpha" />
     <PackageReference Include="App.Metrics.Formatters.Json" Version="2.0.0-alpha" />
+    <PackageReference Include="Butterfly.Client.AspNetCore" Version="0.0.2" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.5.0-rc2" />
-    <PackageReference Include="Ocelot" Version="2.0.10" />
+    <PackageReference Include="Ocelot" Version="2.0.11" />
   </ItemGroup>
   <ItemGroup>
     <Content Update="appsettings.json">
