+++ src/NanoFabric.Ocelot/Program.cs
@@ -1,10 +1,15 @@
-﻿using Microsoft.AspNetCore.Hosting;
+﻿using App.Metrics;
+using Butterfly.Client.AspNetCore;
+using CacheManager.Core;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Configuration;
-using System.IO;
-using Rafty.Infrastructure;
 using NLog.Web;
+using Ocelot.DependencyInjection;
+using Rafty.Infrastructure;
+using System;
+using System.IO;
 
 namespace NanoFabric.Ocelot
 {
@@ -12,13 +17,14 @@ namespace NanoFabric.Ocelot
     {
         public static void Main(string[] args)
         {
-            var hostingconfig = new ConfigurationBuilder()
+            var hostingconfig = new Microsoft.Extensions.Configuration.ConfigurationBuilder()
             .SetBasePath(Directory.GetCurrentDirectory())
             .AddJsonFile("hosting.json", optional: true)
+            .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
             .AddCommandLine(args)
             .Build();
 
-            var url = hostingconfig.GetValue<string>("urls");
+            var url = hostingconfig.GetValue<string>("urls");            
 
             IWebHostBuilder builder = new WebHostBuilder();
             builder.ConfigureServices(s => {
@@ -31,12 +37,50 @@ namespace NanoFabric.Ocelot
                 .ConfigureAppConfiguration((hostingContext, config) =>
                 {
                     config.SetBasePath(hostingContext.HostingEnvironment.ContentRootPath);
-                    var env = hostingContext.HostingEnvironment;
-                    config.AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                        .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true);
+                    var env = hostingContext.HostingEnvironment;                    
                     config.AddJsonFile("configuration.json");
                     config.AddJsonFile("peers.json");
                     config.AddEnvironmentVariables();
+                })
+                .ConfigureServices(services =>
+                {
+                    Action<CacheManager.Core.ConfigurationBuilderCachePart> settings = (x) =>
+                    {
+                        x.WithMicrosoftLogging(log =>
+                        {
+                            log.AddConsole(LogLevel.Debug);
+                        })
+                        .WithDictionaryHandle();
+                    };
+
+                    services.AddAuthentication()
+                        .AddJwtBearer("TestKey", x =>
+                        {
+                            x.Authority = "test";
+                            x.Audience = "test";
+                        });
+
+                    services.AddOcelot()
+                        .AddCacheManager(settings)
+                        .AddAdministration("/administration", "secret")
+                        .AddRafty();
+
+
+                    var metrics = AppMetrics.CreateDefaultBuilder()
+                               .Build();
+
+                    services.AddMetrics(metrics);
+                    services.AddMetricsTrackingMiddleware();
+                    services.AddMetricsEndpoints();
+                    services.AddMetricsReportScheduler();
+
+                    var collectorUrl = hostingconfig.GetValue<string>("Butterfly:CollectorUrl");
+
+                    services.AddButterfly(option =>
+                    {
+                        option.CollectorUrl = collectorUrl;
+                        option.Service = "Ocelot";
+                    });
                 })
                  .ConfigureLogging((hostingContext, logging) =>
                  {
