+++ src/NanoFabric.Ocelot/appsettings.json
@@ -18,6 +18,8 @@
     "MetricsEndpointEnabled": true,
     "MetricsTextEndpointEnabled": true,
     "EnvironmentInfoEndpointEnabled": true
+  },
+  "Butterfly": {
+    "CollectorUrl": "http://localhost:9618"
   }
-
 }
